use different MuNumericTypes for _number and vector
