+++ src/schema/type.ts
@@ -1,13 +1,3 @@
-export type MuNumericType =
-    'float32'   |
-    'float64'   |
-    'int8'      |
-    'int16'     |
-    'int32'     |
-    'uint8'     |
-    'uint16'    |
-    'uint32';
-
 export type MuStringType =
     'ascii'         |
     'fixed-ascii'   |
